pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reel-generator"
version = "0.1.1"
description = "Chat to short vertical video reel generator"
readme = "README.md"
authors = [{name="Your Name"}]
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "pydantic==2.8.2",
    "moviepy==1.0.3",
    "gTTS==2.5.1",
    "python-dotenv==1.0.1",
    "pillow==10.3.0",
    "requests==2.32.3",
    "numpy==1.26.4"
]

[project.optional-dependencies]
dev = ["pytest==8.2.2"]

[tool.pytest.ini_options]
pythonpath = ["src"]
